Added Fdate processing

+++ StartMyGuiTrimmer.R
@@ -1,6 +1,6 @@
 StartMyGUI <- function() {
   mtime=file.mtime(svt)
-  print('Gui Started')
+  print(paste('Gui Started - NEW',mtime))
   startt=EnterStartStop()
   if(.GlobalEnv$convert){
     .GlobalEnv$convert=FALSE
@@ -11,14 +11,15 @@ StartMyGUI <- function() {
                 svt,'" ',of,',' ,sep=''),translate = TRUE)
     #     shell(paste('c:/Users/Larry/Documents/hexDump/bin/converth265.bat "',
     #                 svt,'" c:/Users/Larry/Documents/out.mp4',sep=''),translate = TRUE)
-    if(!file.rename(of,sub('REDUCE','',svt))){
+    ofn=sub('REDUCE','',svt)
+    if(!file.rename(of,ofn)){
       print('file rename back to orig failed - REDUCE')
     }else{
       unlink(of)
       print('file renamed back to orig - REDUCE')
       dx=data.frame(dtn=NA,fn=NA,times=NA)
-      dx$dtn=mtime
-      dx$fn=normalizePath(as.character(svt),winslash = '/')
+      dx$dtn=mtime+(8*3600) #add 8 hours to make GMT
+      dx$fn=normalizePath(as.character(ofn),winslash = '/')
       dx$times=paste('Y:',getYear(dx$dtn),' M:',getMonth(dx$dtn),' D:',getDay(dx$dtn),' H:',as.POSIXlt(dx$dtn)$hour,
                      ' I:',as.POSIXlt(dx$dtn)$min,' S:' ,as.POSIXlt(dx$dtn)$sec,sep='')
       cmd=paste('shell(','"fdate',dx$fn,dx$times,'")')
@@ -26,35 +27,53 @@ StartMyGUI <- function() {
       print('file mtime back to orig - REDUCE')
     }
     startt=NULL
+  }else{
+    if(.GlobalEnv$Fdate){
+      source('~/pllist.git/dfxprocess.R')
+      .GlobalEnv$Fdate=TRUE
+    }
   }
-  
-  print(paste('len startt=',len(startt)))
-  if(len(startt)>0){
-    print(paste('startt=',startt))
-    endtt=0
-    if(!.GlobalEnv$ToEnd)
-      endtt=EnterStartStop("Enter Time Duration (secs) or (mm:ss), 
+  if(!.GlobalEnv$Fdate){  
+    print(paste('len startt=',len(startt)))
+    if(len(startt)>0){
+      print(paste('startt=',startt))
+      endtt=0
+      if(!.GlobalEnv$ToEnd)
+        endtt=EnterStartStop("Enter Time Duration (secs) or (mm:ss), 
                            or Enter/Esc for End of File\n",TRUE)
-    
-    #if(startt > 0 & len(endtt)){
+      
+      #if(startt > 0 & len(endtt)){
       if(len(endtt)){
-      svtt='c:/RealPlayerDownloads/trimmed.mp4'
-      unlink('~/temppt.mp4')
-      unlink(svtt)
-      file.rename(svt,'~/temppt.mp4')
-      if(.GlobalEnv$ToEnd){
-        endtt=10000
-        .GlobalEnv$ToEnd=FALSE
-      }
-      cmdd=paste('shell("ffmpeg.exe -ss',startt,' -i c:/users/Larry/Documents/temppt.mp4 -t',endtt,'-c:v copy -c:a copy',svtt,'",mustWork=NA,translate=TRUE)')
-      print(cmdd)
-      eval(parse(text=cmdd))
-      svt1=sub('TRIM','',svt)
-      svtO=paste(file_path_sans_ext(svt1),'_cut.',file_ext(svt1),sep='') # add _New to original filename
-      file.rename(svtt,svtO) # replace svt has trimmed with start to end
-      file.rename('~/temppt.mp4',svt) # keep original file
-    }else
-      print('Invalid start/end time')
+        svtt='c:/RealPlayerDownloads/trimmed.mp4'
+        unlink('~/temppt.mp4')
+        unlink(svtt)
+        file.rename(svt,'~/temppt.mp4')
+        if(.GlobalEnv$ToEnd){
+          endtt=10000
+          .GlobalEnv$ToEnd=FALSE
+        }
+        cmdd=paste('shell("ffmpeg.exe -ss',startt,' -i c:/users/Larry/Documents/temppt.mp4 -t',endtt,'-c:v copy -c:a copy',svtt,'",mustWork=NA,translate=TRUE)')
+        print(cmdd)
+        eval(parse(text=cmdd))
+        svt1=sub('TRIM','',svt)
+        svtO=paste(file_path_sans_ext(svt1),'_cut.',file_ext(svt1),sep='') # add _New to original filename
+        file.rename(svtt,svtO) # replace svt has trimmed with start to end
+        file.rename('~/temppt.mp4',svt) # keep original file
+        print('file renamed back to orig - REDUCE')
+        
+        dx=data.frame(dtn=NA,fn=NA,times=NA)
+        dx$dtn=mtime+(8*3600) #add 8 hours to make GMT
+        dx$fn=normalizePath(as.character(svtO),winslash = '/')
+        dx$times=paste('Y:',getYear(dx$dtn),' M:',getMonth(dx$dtn),' D:',getDay(dx$dtn),' H:',as.POSIXlt(dx$dtn)$hour,
+                       ' I:',as.POSIXlt(dx$dtn)$min,' S:' ,as.POSIXlt(dx$dtn)$sec,sep='')
+        cmd=paste('shell(','"fdate',dx$fn,dx$times,'")')
+        eval(parse(text=cmd))
+        print('file mtime back to orig - REDUCE')
+        print(file.mtime(svtO))
+        gmessage(paste(cmd,file.mtime(svtO),svtO))
+      }else
+        print('Invalid start/end time')
+    }
+    return()
   }
-  return()
-} 
+}
